refactor: #59 mover helpers a infra/utils/ y DRY en BackendClient
- Mueve backend-client, logger, retry, types de lambdas/helpers/ a infra/utils/
- BackendClient: método request() privado elimina código repetido (150 → 96 líneas)
- Nuevo BackendClientError con statusCode y responseBody
- Interfaces BackendClientConfig y BackendMatch movidas a types.ts
- Actualiza imports en los 4 lambdas
