+++ infra/lambdas/notify-users.ts
@@ -1,7 +1,7 @@
 import type { SQSEvent, SQSRecord } from "aws-lambda";
-import { BackendClient } from "./helpers/backend-client.js";
-import { createEventLogger } from "./helpers/logger.js";
-import { RetryExhaustedError } from "./helpers/retry.js";
+import { BackendClient } from "../utils/backend-client.js";
+import { createEventLogger } from "../utils/logger.js";
+import { RetryExhaustedError } from "../utils/retry.js";
 
 const BACKEND_URL = process.env.BACKEND_URL ?? "";
