+++ infra/utils/types.ts
@@ -1,3 +1,5 @@
+import type { RetryOptions } from "./retry.js";
+
 export interface MatchFinishedDetail {
   match_id: string;
   tournament_id: string;
@@ -11,6 +13,24 @@ export interface CalculateInstanceRankingDetail {
   match_id: string;
   tournament_id: string;
   instance_id: string;
+}
+
+export interface BackendClientConfig {
+  baseUrl: string;
+  retryOptions?: RetryOptions;
+}
+
+export interface BackendMatch {
+  id: string;
+  tournament_id: string;
+  home_team_id: string;
+  away_team_id: string;
+  kick_off: string;
+  home_goals: number;
+  away_goals: number;
+  status: string;
+  stage: string;
+  venue: string;
 }
 
 export type ProcessingStatus = "STARTED" | "RETRYING" | "SUCCESS" | "FAILED";
